requirements.txt
opencv-python~=4.12.0.88
numpy~=2.2.6
torch~=2.5.1+cu118
pillow~=12.0.0
torchvision~=0.20.1+cu118
PyYAML~=6.0.3
tqdm~=4.67.1
matplotlib~=3.10.7
pandas~=2.3.3
seaborn~=0.13.2
ipython~=9.8.0
requests~=2.32.5
pycocotools~=2.0.10
psutil~=7.1.3
thop~=0.1.1-2209072238
ultralytics-thop~=2.0.18
scipy~=1.16.3